refactor(shards): share the font name-table reader
fontVersion was copied byte-for-byte into seven script shards. It moves to
scripts/font-version.ts and is imported via the existing @/ alias.

The helper cannot live under shards/, because the Test workflow derives
package identifiers from the basename of every changed file matching
shards/**, so a module there would be run as a shard and fail.

Verified the alias resolves at runtime under bun, not just for typecheck,
by parsing the Atkinson Next and Fira Sans Extra Condensed fonts through
the imported function.
